requirements.txt
Django==1.11.2
django-extensions==1.7.9
django-filter==1.0.4
djangorestframework==3.6.3
Markdown==2.6.8
olefile==0.44
Pillow==4.1.1
pygraphviz==1.3.1
pytz==2017.2
six==1.10.0
